+++ programs/drift/src/controller/orders.rs
@@ -6,7 +6,6 @@ use std::u64;
 use anchor_lang::prelude::*;
 use solana_program::msg;
 
-use crate::controller;
 use crate::controller::funding::settle_funding_payment;
 use crate::controller::lp::burn_lp_shares;
 use crate::controller::position;
@@ -80,6 +79,7 @@ use crate::validation;
 use crate::validation::order::{
     validate_order, validate_order_for_force_reduce_only, validate_spot_order,
 };
+use crate::{controller, MARGIN_PRECISION};
 
 #[cfg(test)]
 mod tests;
@@ -1026,7 +1026,7 @@ pub fn fill_perp_order(
     let amm_can_skip_duration: bool;
     let amm_lp_allowed_to_jit_make: bool;
     let oracle_valid_for_amm_fill: bool;
-
+    let oracle_stale_for_margin: bool;
     let mut amm_is_available = !state.amm_paused()?;
     {
         let market = &mut perp_market_map.get_ref_mut(&market_index)?;
@@ -1047,6 +1047,12 @@ pub fn fill_perp_order(
 
         oracle_valid_for_amm_fill =
             is_oracle_valid_for_action(_oracle_validity, Some(DriftAction::FillOrderAmm))?;
+
+        oracle_stale_for_margin = oracle_price_data.delay
+            > state
+                .oracle_guard_rails
+                .validity
+                .slots_before_stale_for_margin;
 
         amm_is_available &= oracle_valid_for_amm_fill;
         amm_is_available &= !market.is_operation_paused(PerpOperation::AmmFill);
@@ -1263,6 +1269,7 @@ pub fn fill_perp_order(
         state.min_perp_auction_duration,
         amm_availability,
         fill_mode,
+        oracle_stale_for_margin,
     )?;
 
     if base_asset_amount != 0 {
@@ -1683,6 +1690,7 @@ fn fulfill_perp_order(
     min_auction_duration: u8,
     amm_availability: AMMAvailability,
     fill_mode: FillMode,
+    oracle_stale_for_margin: bool,
 ) -> DriftResult<(u64, u64)> {
     let market_index = user.orders[user_order_index].market_index;
 
@@ -1697,6 +1705,7 @@ fn fulfill_perp_order(
         perp_market.amm.order_tick_size,
         perp_market.is_prediction_market(),
     )?;
+    let perp_market_oi_before = perp_market.get_open_interest();
     drop(perp_market);
 
     let fulfillment_methods = {
@@ -1874,19 +1883,25 @@ fn fulfill_perp_order(
             -(base_asset_amount as i64)
         };
 
+        let mut context = MarginContext::standard(if user_order_position_decreasing {
+            MarginRequirementType::Maintenance
+        } else {
+            MarginRequirementType::Fill
+        })
+        .fuel_perp_delta(market_index, taker_base_asset_amount_delta)
+        .fuel_numerator(user, now);
+
+        if oracle_stale_for_margin && !user_order_position_decreasing {
+            context = context.margin_ratio_override(MARGIN_PRECISION);
+        }
+
         let taker_margin_calculation =
             calculate_margin_requirement_and_total_collateral_and_liability_info(
                 user,
                 perp_market_map,
                 spot_market_map,
                 oracle_map,
-                MarginContext::standard(if user_order_position_decreasing {
-                    MarginRequirementType::Maintenance
-                } else {
-                    MarginRequirementType::Fill
-                })
-                .fuel_perp_delta(market_index, taker_base_asset_amount_delta)
-                .fuel_numerator(user, now),
+                context,
             )?;
 
         user_stats.update_fuel_bonus(
@@ -1916,11 +1931,27 @@ fn fulfill_perp_order(
             Some(makers_and_referrer_stats.get_ref_mut(&maker.authority)?)
         };
 
-        let margin_type = select_margin_type_for_perp_maker(
+        let (margin_type, maker_risk_increasing) = select_margin_type_for_perp_maker(
             &maker,
             maker_base_asset_amount_filled,
             market_index,
         )?;
+
+        let mut context = MarginContext::standard(margin_type)
+            .fuel_perp_delta(market_index, -maker_base_asset_amount_filled)
+            .fuel_numerator(&maker, now);
+
+        if oracle_stale_for_margin {
+            validate!(
+                user_order_position_decreasing || !maker_risk_increasing,
+                ErrorCode::InvalidOracle,
+                "taker or maker must be reducing position if oracle stale for margin"
+            )?;
+
+            if maker_risk_increasing {
+                context = context.margin_ratio_override(MARGIN_PRECISION);
+            }
+        }
 
         let maker_margin_calculation =
             calculate_margin_requirement_and_total_collateral_and_liability_info(
@@ -1928,9 +1959,7 @@ fn fulfill_perp_order(
                 perp_market_map,
                 spot_market_map,
                 oracle_map,
-                MarginContext::standard(margin_type)
-                    .fuel_perp_delta(market_index, -maker_base_asset_amount_filled)
-                    .fuel_numerator(&maker, now),
+                context,
             )?;
 
         if let Some(mut maker_stats) = maker_stats {
@@ -1952,6 +1981,15 @@ fn fulfill_perp_order(
             );
             return Err(ErrorCode::InsufficientCollateral);
         }
+    }
+
+    if oracle_stale_for_margin {
+        let perp_market_oi_after = perp_market_map.get_ref(&market_index)?.get_open_interest();
+        validate!(
+            perp_market_oi_after <= perp_market_oi_before,
+            ErrorCode::InvalidOracle,
+            "oracle stale for margin but open interest increased"
+        )?;
     }
 
     Ok((base_asset_amount, quote_asset_amount))
@@ -3746,6 +3784,7 @@ pub fn fill_spot_order(
         slot,
     )?;
 
+    let mut oracle_stale_for_margin = false;
     {
         let mut quote_market = spot_market_map.get_quote_spot_market_mut()?;
         let oracle_price_data = oracle_map.get_price_data(&quote_market.oracle_id())?;
@@ -3754,6 +3793,12 @@ pub fn fill_spot_order(
         let mut base_market = spot_market_map.get_ref_mut(&order_market_index)?;
         let oracle_price_data = oracle_map.get_price_data(&base_market.oracle_id())?;
         update_spot_market_cumulative_interest(&mut base_market, Some(oracle_price_data), now)?;
+
+        oracle_stale_for_margin = oracle_price_data.delay
+            > state
+                .oracle_guard_rails
+                .validity
+                .slots_before_stale_for_margin;
 
         fulfillment_params.validate_markets(&base_market, &quote_market)?;
 
@@ -3858,6 +3903,7 @@ pub fn fill_spot_order(
         slot,
         &state.spot_fee_structure,
         fulfillment_params,
+        oracle_stale_for_margin,
     )?;
 
     if base_asset_amount != 0 {
@@ -4121,6 +4167,7 @@ fn fulfill_spot_order(
     slot: u64,
     fee_structure: &FeeStructure,
     fulfillment_params: &mut dyn SpotFulfillmentParams,
+    oracle_stale_for_margin: bool,
 ) -> DriftResult<(u64, u64)> {
     let base_market_index = user.orders[user_order_index].market_index;
     let order_direction = user.orders[user_order_index].direction;
@@ -4303,7 +4350,7 @@ fn fulfill_spot_order(
         true,
     );
 
-    let margin_type = if order_direction == PositionDirection::Long {
+    let (margin_type, taker_risk_increasing) = if order_direction == PositionDirection::Long {
         // sell quote, buy base
         select_margin_type_for_swap(
             &quote_market,
@@ -4334,24 +4381,30 @@ fn fulfill_spot_order(
     drop(base_market);
     drop(quote_market);
 
+    let mut context = MarginContext::standard(margin_type)
+        .fuel_spot_deltas([
+            (
+                base_market_index,
+                base_token_amount_before.safe_sub(base_token_amount_after)?,
+            ),
+            (
+                QUOTE_SPOT_MARKET_INDEX,
+                quote_token_amount_before.safe_sub(quote_token_amount_after)?,
+            ),
+        ])
+        .fuel_numerator(user, now);
+
+    if oracle_stale_for_margin && taker_risk_increasing {
+        context = context.margin_ratio_override(MARGIN_PRECISION);
+    }
+
     let taker_margin_calculation =
         calculate_margin_requirement_and_total_collateral_and_liability_info(
             user,
             perp_market_map,
             spot_market_map,
             oracle_map,
-            MarginContext::standard(margin_type)
-                .fuel_spot_deltas([
-                    (
-                        base_market_index,
-                        base_token_amount_before.safe_sub(base_token_amount_after)?,
-                    ),
-                    (
-                        QUOTE_SPOT_MARKET_INDEX,
-                        quote_token_amount_before.safe_sub(quote_token_amount_after)?,
-                    ),
-                ])
-                .fuel_numerator(user, now),
+            context,
         )?;
 
     // user hasnt recieved initial fuel or below global start time
@@ -4393,7 +4446,7 @@ fn fulfill_spot_order(
             .get_spot_position(base_market_index)?
             .get_signed_token_amount(&base_market)?;
 
-        let margin_type = if maker_direction == PositionDirection::Long {
+        let (margin_type, maker_risk_increasing) = if maker_direction == PositionDirection::Long {
             // sell quote, buy base
             select_margin_type_for_swap(
                 &quote_market,
@@ -4424,26 +4477,38 @@ fn fulfill_spot_order(
         drop(base_market);
         drop(quote_market);
 
+        let mut context = MarginContext::standard(margin_type)
+            .fuel_spot_deltas([
+                (
+                    base_market_index,
+                    maker_base_token_amount_before.safe_sub(maker_base_token_amount_after)?,
+                ),
+                (
+                    QUOTE_SPOT_MARKET_INDEX,
+                    maker_quote_token_amount_before.safe_sub(maker_quote_token_amount_after)?,
+                ),
+            ])
+            .fuel_numerator(&maker, now);
+
+        if oracle_stale_for_margin {
+            validate!(
+                !maker_risk_increasing || !taker_risk_increasing,
+                ErrorCode::InvalidOracle,
+                "maker or taker must be reducing position if oracle stale for margin"
+            )?;
+
+            if maker_risk_increasing {
+                context = context.margin_ratio_override(MARGIN_PRECISION);
+            }
+        }
+
         let maker_margin_calculation: MarginCalculation =
             calculate_margin_requirement_and_total_collateral_and_liability_info(
                 &maker,
                 perp_market_map,
                 spot_market_map,
                 oracle_map,
-                MarginContext::standard(margin_type)
-                    .fuel_spot_deltas([
-                        (
-                            base_market_index,
-                            maker_base_token_amount_before
-                                .safe_sub(maker_base_token_amount_after)?,
-                        ),
-                        (
-                            QUOTE_SPOT_MARKET_INDEX,
-                            maker_quote_token_amount_before
-                                .safe_sub(maker_quote_token_amount_after)?,
-                        ),
-                    ])
-                    .fuel_numerator(&maker, now),
+                context,
             )?;
 
         if let Some(mut maker_stats) = maker_stats {